gui/dbSta: break the dependency cycle
Move the power density heatmap source from dbSta to gui.  Remove the
obsolete highlight_path command (timing path view replaces it).  Gui
depends on dbSta but not visa-versa.
