+++ src/dbSta/include/db_sta/dbSta.hh
@@ -27,8 +27,6 @@ class dbNetwork;
 class dbStaReport;
 class dbStaHistogram;
 class dbStaCbk;
-class AbstractPathRenderer;
-class AbstractPowerDensityDataSource;
 class PatternMatch;
 class TestCell;
 
@@ -123,10 +121,6 @@ class dbSta : public Sta, public odb::dbDatabaseObserver
                 odb::dbDatabase* db,
                 utl::Logger* logger);
 
-  void setPathRenderer(std::unique_ptr<AbstractPathRenderer> path_renderer);
-  void setPowerDensityDataSource(std::unique_ptr<AbstractPowerDensityDataSource>
-                                     power_density_data_source);
-
   // Creates a dbSta instance for the given dbBlock using the same context as
   // this dbSta instance (e.g. TCL interpreter, units, etc.)
   std::unique_ptr<dbSta> makeBlockSta(odb::dbBlock* block);
@@ -155,9 +149,6 @@ class dbSta : public Sta, public odb::dbDatabaseObserver
   void updateComponentsState() override;
   void registerStaState(dbStaState* state);
   void unregisterStaState(dbStaState* state);
-
-  // Highlight path in the gui.
-  void highlight(Path* path);
 
   std::string getInstanceTypeText(InstType type) const;
   InstType getInstanceType(odb::dbInst* inst);
@@ -204,9 +195,6 @@ class dbSta : public Sta, public odb::dbDatabaseObserver
   std::set<dbStaState*> sta_states_;
 
   std::unique_ptr<BufferUseAnalyser> buffer_use_analyser_;
-
-  std::unique_ptr<AbstractPathRenderer> path_renderer_;
-  std::unique_ptr<AbstractPowerDensityDataSource> power_density_data_source_;
 };
 
 // Utilities for TestCell
