+++ src/dbSta/src/MakeDbSta.cc
@@ -7,11 +7,8 @@
 
 #include <memory>
 
-#include "PathRenderer.h"
 #include "db_sta/dbNetwork.hh"
 #include "db_sta/dbSta.hh"
-#include "gui/gui.h"
-#include "heatMap.h"
 #include "odb/db.h"
 #include "ord/OpenRoad.hh"
 #include "utl/decode.h"
@@ -42,12 +39,6 @@ void initDbSta(OpenRoad* openroad)
   utl::Logger* logger = openroad->getLogger();
   sta->initVars(openroad->tclInterp(), openroad->getDb(), logger);
   sta::Sta::setSta(sta);
-
-  if (gui::Gui::enabled()) {
-    sta->setPathRenderer(std::make_unique<sta::PathRenderer>(sta));
-  }
-  sta->setPowerDensityDataSource(
-      std::make_unique<sta::PowerDensityDataSource>(sta, logger));
 
   Tcl_Interp* tcl_interp = openroad->tclInterp();
