+++ src/gui/include/gui/MakeGui.h
@@ -3,6 +3,7 @@
 
 #include <tcl.h>
 
+#include "db_sta/dbSta.hh"
 #include "odb/db.h"
 #include "utl/Logger.h"
 
@@ -11,6 +12,9 @@ namespace ord {
 // There is no make/delete GUI as it is created at startup and can't
 // be deleted.
 
-void initGui(Tcl_Interp* interp, odb::dbDatabase* db, utl::Logger* logger);
+void initGui(Tcl_Interp* interp,
+             odb::dbDatabase* db,
+             sta::dbSta* sta,
+             utl::Logger* logger);
 
 }  // namespace ord
