+++ src/gui/include/gui/gui.h
@@ -25,6 +25,10 @@
 
 struct Tcl_Interp;
 
+namespace sta {
+class dbSta;
+}
+
 namespace utl {
 class Logger;
 }  // namespace utl
@@ -33,6 +37,7 @@ namespace gui {
 class HeatMapDataSource;
 class PinDensityDataSource;
 class PlacementDensityDataSource;
+class PowerDensityDataSource;
 class Painter;
 class Selected;
 class Options;
@@ -779,7 +784,7 @@ class Gui
   static bool enabled();
 
   // initialize the GUI
-  void init(odb::dbDatabase* db, utl::Logger* logger);
+  void init(odb::dbDatabase* db, sta::dbSta* sta, utl::Logger* logger);
 
  private:
   Gui();
@@ -837,6 +842,7 @@ class Gui
 
   std::unique_ptr<PinDensityDataSource> pin_density_heat_map_;
   std::unique_ptr<PlacementDensityDataSource> placement_density_heat_map_;
+  std::unique_ptr<PowerDensityDataSource> power_density_heat_map_;
 
   static Gui* singleton_;
